Implement input output channels
Note: Doing this I noticed that my debugger pausing isn't really working
Instead the UI is just not being updated while the machine runs in the
background. To prevent this I really need to implement CPU state that
will enable skipping "next_instruction" calls and halting the CPU.

To do that I'm thinking about moving the HaltedState from registers
to the CPU. This makes sense given halted states only get set on the
CPU. Doing that I should be able to make an Enum that contains both
HaltedState and DebugState and should make life easier?